Task3_ImageCaptioning/requirements.txt
tensorflow==2.12.0
numpy==1.23.5
pillow
opencv-python
